rouze-signal-miner/requirements.txt
praw==7.7.1
PyYAML==6.0.2
pandas==2.2.2
feedparser==6.0.11
python-dateutil==2.9.0.post0
tenacity==9.0.0
python-dotenv==1.0.1
